Issue #472: API: locales service returns different structure per repo
The reason is that we sure asort() which sorts arrays taking keys into account, if the source file is already sorted correctly, asort has no effect and the data returned is a simple array. If asort() changed the order of the locales, the returned array now has numeric keys. We need to use sort().